Remove special handling for ReleaseName
Previously, a `<Spec.ReleaseNamespace>-` prefix was added to `Spec.ReleaseName`,
but only when `Spec.ReleaseNamespace` was set. This conditional prefixing seems
confusing and also doesn't seem to accomplish the presumptive purpose of avoiding
release conflicts. For example, if HelmReleases in two different namespaces with
the same name set the same `Spec.ReleaseNamespace`, we will have a conflict.

This instead just defaults `Spec.ReleaseNamespace` to the HelmRelease namespace,
which matches the release name handling, and leaves any conflict resolution
up to the user.