Increase default terminationGracePeriodSeconds
To a total of 10 minutes (600s), this allows most release process that
just have been started and make use of the default timeouts to finish.

In a future release, we will likely want to be able to stop a
reconciliation process between the reconciliation steps and continue on
the pod restart (e.g. upgrade -> terminate -> start -> continue with
test).

Signed-off-by: Hidde Beydals <hello@hidde.co>