api: allow configuration of `h` unit for timeouts
Unlike the other Flux components, there are some scenarios in which a
timeout of 1h (or more) is justified to be able to Helm actions which
take a long time to finish.

To support this, widen the validation added in
54aee36b3a9adaade8625965604b5d941f63dcd3 to allow the configuration
of the `h` time unit again. At the cost of theoretically keeping the
workers busy for a long duration of time (and thereby not processing
other changes), but without causing a full denial of service.

Signed-off-by: Hidde Beydals <hello@hidde.co>