Allow opt-out of drift correction
This enhances the drift detection to allow detection without acting on
it by adding a `CorrectDrift` feature gate. When set to `false` (i.e.
`--feature-gates=DetectDrift=true,CorrectDrift=false`), the controller
will only log the difference it detected without acting on it.

This allows for a smoother transition to drift detection, as issues
with a number of existing releases can be solved before the feature
is enabled in full.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>