reconcile: allow unlock without current
As persisting to the API server before an unexpected termination is not
feasible, allow a release to be unlocked without it being observed.

This has the potential result of interfering with a user running `helm`
by hand, but the further process should be capable of dealing with any
unexpected failure resulting from this.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>