ci: increase timeout delete-ns tests
This test has shown to be a flake on multiple ocassions.

After running it locally a dozen of times, my conclusion is that the
time it takes for Kubernetes to shutdown a Pod may sometimes be a tiny
bit longer than anticipated by the current timeout.

Therefore, increase it to 5 minutes and provide more context on the
state of the namespace in case a timeout occurs.

Signed-off-by: Hidde Beydals <hidde@hhh.computer>